Extract editor active selection helpers
