+++ CMakeLists.txt
@@ -160,6 +160,7 @@ target_sources(
         src/game_data_conditions_presentation.c
         src/game_data_controller_runtime.c
         src/game_data_controller_binding_helpers.c
+        src/game_data_editor_active_selection.c
         src/game_data_editor_command_preview.c
         src/game_data_editor_command_transactions.c
         src/game_data_editor_output_actions.c
